requirements.txt
Requirements for the SEO Blog Generator

Below are the detailed requirements for the SEO Blog Generator assignment. These requirements are divided into functional, non-functional, and technical aspects to ensure the system meets its goals effectively.

Functional Requirements
Research Agent:

For Reasearch agent i use web Scraping the LinkedIn Talent Blog or similar websites to identify trending HR topics.

Collect relevant information about the selected topic to use as input for the blog post.

Handle errors such as website structure changes, network issues, or missing data gracefully.

Content Planning Agent:

Creating a structured outline for the blog post, including:

A catchy title.

An introduction to the topic.

Main sections (e.g., trends, insights, or case studies).

A conclusion summarizing the key points.

Ensure the outline is flexible enough to adapt to different topics.

Content Generation Agent:

For content Generation Using the GPT-2 model to generate detailed and coherent content for each section of the outline.

Ensure the generated content is relevant, informative, and free from grammatical errors.

Provide fallback content if GPT-2 fails to generate meaningful text.

SEO Optimization Agent:

Optimize the blog content for search engines by:

Adding relevant keywords (e.g., "HR trends 2023").

Using proper HTML tags (e.g., <h1>, <h2>, <strong>).

Ensuring the content is well-structured and easy to read.

Including meta elements like <title> and <meta name="description"> in the final HTML output.

Review Agent:

Proofread the generated content to ensure it is grammatically correct and readable.

Using NLTK  similar tools for sentence tokenization and basic grammar checks.

Main System:

Integrate all agents into a seamless workflow.

Saving the final blog post as an HTML file (hr_trends_blog.html).

Providing clear logs and status updates during the execution process.

Non-Functional Requirements
Performance:

The system  generate a blog post within a reasonable time frame (e.g., under 5 minutes).

Handle large amounts of text data efficiently without significant delays.

Scalability:

The system  be modular and extensible, allowing for the addition of new agents or features in the future.

Reliability:

The system should handle errors gracefully, such as failed web requests or GPT-2 generation issues.

Provide fallback mechanisms for critical failures (e.g., default content if GPT-2 fails).

Usability:

The system should be easy to set up and run with minimal configuration.

Providing clear and concise documentation for installation and execution.

SEO Compliance:

The generated blog post should follow SEO best practices, including:

Proper use of headings (<h1>, <h2>, etc.).

Keyword optimization.

Meta tags and descriptions.

Technical Requirements

Programming Language:

Python 3.7 or higher.

Libraries and Frameworks Used In Project:

Requests: For making HTTP requests to fetch web content.

BeautifulSoup: For parsing HTML and extracting data.

Transformers (Hugging Face): For using the GPT-2 model for text generation.

PyTorch: For running the GPT-2 model.

NLTK: For text processing and sentence tokenization.

Data Sources:

I used LinkedIn Talent Blog webpage for data source

Output:

The final blog post is saved as  HTML file (hr_trends_blog.html).

Error Handling:

Implement robust error handling for:

Web scraping failures.

GPT-2 content generation failures.

Delivered Code

Source Code:

Python script(s) implementing the SEO Blog Generator.

Output:

Generated blog post in HTML format (hr_trends_blog.html).

Input:

The system scrapes the LinkedIn Talent Blog for trending HR topics.

Process:

Research Agent identifies a topic (e.g., "Remote Work in 2023").

Content Planning Agent creates an outline.

Content Generation Agent writes the blog post using GPT-2.

SEO Optimization Agent enhances the content for SEO.

Review Agent proofreads the content.

Output:

A well-structured, SEO-optimized blog post saved as hr_trends_blog.html.

By meeting these requirements, the SEO Blog Generator will be a robust, scalable, and efficient system for generating high-quality, SEO-optimized blog posts. Good luck! 🚀

